build scripts: improve start menu layout
Former-commit-id: 208d53f2c937c86608a1cf63515dbeab736eda04 [formerly 0adff24af87ad11806046fa6c4c80f522ee1ba9e]
Former-commit-id: c619694c4c39f584f79dc70261c48b6282b048e4
Former-commit-id: 73e9c6b17a085ebf97cf18a5cbef5d43adfb3ad4

+++ Utilities/Scripts/bundle_setup/wrapup_fds_install.bat
@@ -27,12 +27,21 @@ assoc .smv=smvDoc>Nul
 
 echo. 
 echo Adding FDS and Smokeview shortcuts to the Start menu.
-rmdir /q /s "%USERPROFILE%\Start Menu\Programs\FDS5"
+if exist "%USERPROFILE%\Start Menu\Programs\FDS5" rmdir /q /s "%USERPROFILE%\Start Menu\Programs\FDS5"
+
 mkdir "%USERPROFILE%\Start Menu\Programs\FDS5"
-"%CD%\shortcut.exe" /F:"%USERPROFILE%\Start Menu\Programs\FDS5\Guides_and_Release_Notes.lnk"  /T:"%CD%\Documentation\Guides_and_Release_Notes" /A:C >NUL
-"%CD%\shortcut.exe" /F:"%USERPROFILE%\Start Menu\Programs\FDS5\FDS_on_the_Web.lnk"  /T:"%CD%\Documentation\FDS_on_the_Web" /A:C >NUL
 "%CD%\shortcut.exe" /F:"%USERPROFILE%\Start Menu\Programs\FDS5\Overview.lnk"  /T:"%CD%\Documentation\Overview.html" /A:C >NUL
-"%CD%\shortcut.exe" /F:"%USERPROFILE%\Start Menu\Programs\FDS5\Updates.lnk"  /T:"%CD%\Documentation\FDS_on_the_Web/Updates.url" /A:C >NUL
+copy "%CD%\Documentation\FDS_on_the_Web\Updates.url" "%USERPROFILE%\Start Menu\Programs\FDS5\Updates.url"  
+
+mkdir "%USERPROFILE%\Start Menu\Programs\FDS5\FDS_on_the_Web"
+copy "%CD%\Documentation\FDS_on_the_Web\D*"          "%USERPROFILE%\Start Menu\Programs\FDS5\FDS_on_the_Web"
+copy "%CD%\Documentation\FDS_on_the_Web\O*"          "%USERPROFILE%\Start Menu\Programs\FDS5\FDS_on_the_Web"
+
+mkdir "%USERPROFILE%\Start Menu\Programs\FDS5\User_Guides_and_Release_Notes"
+"%CD%\shortcut.exe" /F:"%USERPROFILE%\Start Menu\Programs\FDS5\User_Guides_and_Release_Notes\FDS_5_User_Guide.lnk"  /T:"%CD%\Documentation\Guides_and_Release_Notes\FDS_5_User_Guide.pdf" /A:C >NUL
+"%CD%\shortcut.exe" /F:"%USERPROFILE%\Start Menu\Programs\FDS5\User_Guides_and_Release_Notes\SMV_5_User_Guide.lnk"  /T:"%CD%\Documentation\Guides_and_Release_Notes\SMV_5_User_Guide.pdf" /A:C >NUL
+"%CD%\shortcut.exe" /F:"%USERPROFILE%\Start Menu\Programs\FDS5\User_Guides_and_Release_Notes\FDS_Release_Notes.lnk"  /T:"%CD%\Documentation\Guides_and_Release_Notes\FDS_Release_Notes.htm" /A:C >NUL
+"%CD%\shortcut.exe" /F:"%USERPROFILE%\Start Menu\Programs\FDS5\User_Guides_and_Release_Notes\Smokeview_release_notes.lnk"  /T:"%CD%\Documentation\Guides_and_Release_Notes\Smokeview_release_notes.pdf" /A:C >NUL
 
 echo.
 echo Adding the directory %CD% to the user path variable for: %USERNAME%
